+++ apps/mobile/CHANGELOG.md
@@ -5,6 +5,20 @@
 项目的 CHANGELOG 格式符合 [Keep a Changelog]，
 且版本号遵循 [Semantic Versioning]。 ~~(然而，事实上遵循的是 [Pride Versioning])~~
 
+## [UNRELEASED]
+
+### Changed
+
+- Orpheus: 优化歌词系统架构
+
+### Added
+
+- 支持车载歌词（Android：通过把当前歌词写入 MediaMetadata.title，在蓝牙 AVRCP 车机上显示）
+
+### Fixed
+
+- 修复因为开发者脑子进水导致的又一次无法上传播放记录的问题
+
 ## [2.4.4] - 2026-04-18
 
 ### Fixed